+++ web/service/outbound/outbound.go
@@ -1,4 +1,4 @@
-package service
+package outbound
 
 import (
 	"encoding/json"
